requirements.txt
telethon==1.35.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3